Factor the wind model into a separate file, winds.h
